Move the formatting flow for talk page updates into a separate class
